- Add slider and category state management, update routing and UI components
