+++ lib/features/home/logic/home_state.freezed.dart
@@ -19,15 +19,6 @@ mixin _$HomeState<T> {
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
     required TResult Function() getHomeProductsLoading,
     required TResult Function(T data) getHomeProductsSuccess,
     required TResult Function(String errorMessage) getHomeProductsFailure,
@@ -36,15 +27,6 @@ mixin _$HomeState<T> {
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
     TResult? Function()? getHomeProductsLoading,
     TResult? Function(T data)? getHomeProductsSuccess,
     TResult? Function(String errorMessage)? getHomeProductsFailure,
@@ -53,15 +35,6 @@ mixin _$HomeState<T> {
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
     TResult Function()? getHomeProductsLoading,
     TResult Function(T data)? getHomeProductsSuccess,
     TResult Function(String errorMessage)? getHomeProductsFailure,
@@ -71,21 +44,6 @@ mixin _$HomeState<T> {
   @optionalTypeArgs
   TResult map<TResult extends Object?>({
     required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
     required TResult Function(GetHomeProductLoading<T> value)
         getHomeProductsLoading,
     required TResult Function(GetHomeProductSuccess<T> value)
@@ -97,18 +55,6 @@ mixin _$HomeState<T> {
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>({
     TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -117,18 +63,6 @@ mixin _$HomeState<T> {
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -200,15 +134,6 @@ class _$HomeInitialImpl<T> implements HomeInitial<T> {
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
     required TResult Function() getHomeProductsLoading,
     required TResult Function(T data) getHomeProductsSuccess,
     required TResult Function(String errorMessage) getHomeProductsFailure,
@@ -220,15 +145,6 @@ class _$HomeInitialImpl<T> implements HomeInitial<T> {
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
     TResult? Function()? getHomeProductsLoading,
     TResult? Function(T data)? getHomeProductsSuccess,
     TResult? Function(String errorMessage)? getHomeProductsFailure,
@@ -240,15 +156,6 @@ class _$HomeInitialImpl<T> implements HomeInitial<T> {
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
     TResult Function()? getHomeProductsLoading,
     TResult Function(T data)? getHomeProductsSuccess,
     TResult Function(String errorMessage)? getHomeProductsFailure,
@@ -264,21 +171,6 @@ class _$HomeInitialImpl<T> implements HomeInitial<T> {
   @optionalTypeArgs
   TResult map<TResult extends Object?>({
     required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
     required TResult Function(GetHomeProductLoading<T> value)
         getHomeProductsLoading,
     required TResult Function(GetHomeProductSuccess<T> value)
@@ -293,18 +185,6 @@ class _$HomeInitialImpl<T> implements HomeInitial<T> {
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>({
     TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -316,18 +196,6 @@ class _$HomeInitialImpl<T> implements HomeInitial<T> {
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -342,1926 +210,6 @@ class _$HomeInitialImpl<T> implements HomeInitial<T> {
 
 abstract class HomeInitial<T> implements HomeState<T> {
   const factory HomeInitial() = _$HomeInitialImpl<T>;
-}
-
-/// @nodoc
-abstract class _$$SliderLoadingImplCopyWith<T, $Res> {
-  factory _$$SliderLoadingImplCopyWith(_$SliderLoadingImpl<T> value,
-          $Res Function(_$SliderLoadingImpl<T>) then) =
-      __$$SliderLoadingImplCopyWithImpl<T, $Res>;
-}
-
-/// @nodoc
-class __$$SliderLoadingImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$SliderLoadingImpl<T>>
-    implements _$$SliderLoadingImplCopyWith<T, $Res> {
-  __$$SliderLoadingImplCopyWithImpl(_$SliderLoadingImpl<T> _value,
-      $Res Function(_$SliderLoadingImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-}
-
-/// @nodoc
-
-class _$SliderLoadingImpl<T> implements SliderLoading<T> {
-  const _$SliderLoadingImpl();
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.sliderLoading()';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$SliderLoadingImpl<T>);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return sliderLoading();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return sliderLoading?.call();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (sliderLoading != null) {
-      return sliderLoading();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return sliderLoading(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return sliderLoading?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (sliderLoading != null) {
-      return sliderLoading(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class SliderLoading<T> implements HomeState<T> {
-  const factory SliderLoading() = _$SliderLoadingImpl<T>;
-}
-
-/// @nodoc
-abstract class _$$SliderSuccessImplCopyWith<T, $Res> {
-  factory _$$SliderSuccessImplCopyWith(_$SliderSuccessImpl<T> value,
-          $Res Function(_$SliderSuccessImpl<T>) then) =
-      __$$SliderSuccessImplCopyWithImpl<T, $Res>;
-  @useResult
-  $Res call({T data});
-}
-
-/// @nodoc
-class __$$SliderSuccessImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$SliderSuccessImpl<T>>
-    implements _$$SliderSuccessImplCopyWith<T, $Res> {
-  __$$SliderSuccessImplCopyWithImpl(_$SliderSuccessImpl<T> _value,
-      $Res Function(_$SliderSuccessImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? data = freezed,
-  }) {
-    return _then(_$SliderSuccessImpl<T>(
-      freezed == data
-          ? _value.data
-          : data // ignore: cast_nullable_to_non_nullable
-              as T,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$SliderSuccessImpl<T> implements SliderSuccess<T> {
-  const _$SliderSuccessImpl(this.data);
-
-  @override
-  final T data;
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.sliderSuccess(data: $data)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$SliderSuccessImpl<T> &&
-            const DeepCollectionEquality().equals(other.data, data));
-  }
-
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, const DeepCollectionEquality().hash(data));
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$SliderSuccessImplCopyWith<T, _$SliderSuccessImpl<T>> get copyWith =>
-      __$$SliderSuccessImplCopyWithImpl<T, _$SliderSuccessImpl<T>>(
-          this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return sliderSuccess(data);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return sliderSuccess?.call(data);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (sliderSuccess != null) {
-      return sliderSuccess(data);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return sliderSuccess(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return sliderSuccess?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (sliderSuccess != null) {
-      return sliderSuccess(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class SliderSuccess<T> implements HomeState<T> {
-  const factory SliderSuccess(final T data) = _$SliderSuccessImpl<T>;
-
-  T get data;
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$SliderSuccessImplCopyWith<T, _$SliderSuccessImpl<T>> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class _$$SliderFailureImplCopyWith<T, $Res> {
-  factory _$$SliderFailureImplCopyWith(_$SliderFailureImpl<T> value,
-          $Res Function(_$SliderFailureImpl<T>) then) =
-      __$$SliderFailureImplCopyWithImpl<T, $Res>;
-  @useResult
-  $Res call({String errorMessage});
-}
-
-/// @nodoc
-class __$$SliderFailureImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$SliderFailureImpl<T>>
-    implements _$$SliderFailureImplCopyWith<T, $Res> {
-  __$$SliderFailureImplCopyWithImpl(_$SliderFailureImpl<T> _value,
-      $Res Function(_$SliderFailureImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? errorMessage = null,
-  }) {
-    return _then(_$SliderFailureImpl<T>(
-      null == errorMessage
-          ? _value.errorMessage
-          : errorMessage // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$SliderFailureImpl<T> implements SliderFailure<T> {
-  const _$SliderFailureImpl(this.errorMessage);
-
-  @override
-  final String errorMessage;
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.sliderFailure(errorMessage: $errorMessage)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$SliderFailureImpl<T> &&
-            (identical(other.errorMessage, errorMessage) ||
-                other.errorMessage == errorMessage));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, errorMessage);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$SliderFailureImplCopyWith<T, _$SliderFailureImpl<T>> get copyWith =>
-      __$$SliderFailureImplCopyWithImpl<T, _$SliderFailureImpl<T>>(
-          this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return sliderFailure(errorMessage);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return sliderFailure?.call(errorMessage);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (sliderFailure != null) {
-      return sliderFailure(errorMessage);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return sliderFailure(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return sliderFailure?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (sliderFailure != null) {
-      return sliderFailure(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class SliderFailure<T> implements HomeState<T> {
-  const factory SliderFailure(final String errorMessage) =
-      _$SliderFailureImpl<T>;
-
-  String get errorMessage;
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$SliderFailureImplCopyWith<T, _$SliderFailureImpl<T>> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class _$$GetCategoriesLoadingImplCopyWith<T, $Res> {
-  factory _$$GetCategoriesLoadingImplCopyWith(
-          _$GetCategoriesLoadingImpl<T> value,
-          $Res Function(_$GetCategoriesLoadingImpl<T>) then) =
-      __$$GetCategoriesLoadingImplCopyWithImpl<T, $Res>;
-}
-
-/// @nodoc
-class __$$GetCategoriesLoadingImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$GetCategoriesLoadingImpl<T>>
-    implements _$$GetCategoriesLoadingImplCopyWith<T, $Res> {
-  __$$GetCategoriesLoadingImplCopyWithImpl(_$GetCategoriesLoadingImpl<T> _value,
-      $Res Function(_$GetCategoriesLoadingImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-}
-
-/// @nodoc
-
-class _$GetCategoriesLoadingImpl<T> implements GetCategoriesLoading<T> {
-  const _$GetCategoriesLoadingImpl();
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.getCategoriesLoading()';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$GetCategoriesLoadingImpl<T>);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return getCategoriesLoading();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return getCategoriesLoading?.call();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoriesLoading != null) {
-      return getCategoriesLoading();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return getCategoriesLoading(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return getCategoriesLoading?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoriesLoading != null) {
-      return getCategoriesLoading(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class GetCategoriesLoading<T> implements HomeState<T> {
-  const factory GetCategoriesLoading() = _$GetCategoriesLoadingImpl<T>;
-}
-
-/// @nodoc
-abstract class _$$GetCategoriesSuccessImplCopyWith<T, $Res> {
-  factory _$$GetCategoriesSuccessImplCopyWith(
-          _$GetCategoriesSuccessImpl<T> value,
-          $Res Function(_$GetCategoriesSuccessImpl<T>) then) =
-      __$$GetCategoriesSuccessImplCopyWithImpl<T, $Res>;
-  @useResult
-  $Res call({T data});
-}
-
-/// @nodoc
-class __$$GetCategoriesSuccessImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$GetCategoriesSuccessImpl<T>>
-    implements _$$GetCategoriesSuccessImplCopyWith<T, $Res> {
-  __$$GetCategoriesSuccessImplCopyWithImpl(_$GetCategoriesSuccessImpl<T> _value,
-      $Res Function(_$GetCategoriesSuccessImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? data = freezed,
-  }) {
-    return _then(_$GetCategoriesSuccessImpl<T>(
-      freezed == data
-          ? _value.data
-          : data // ignore: cast_nullable_to_non_nullable
-              as T,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$GetCategoriesSuccessImpl<T> implements GetCategoriesSuccess<T> {
-  const _$GetCategoriesSuccessImpl(this.data);
-
-  @override
-  final T data;
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.getCategoriesSuccess(data: $data)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$GetCategoriesSuccessImpl<T> &&
-            const DeepCollectionEquality().equals(other.data, data));
-  }
-
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, const DeepCollectionEquality().hash(data));
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$GetCategoriesSuccessImplCopyWith<T, _$GetCategoriesSuccessImpl<T>>
-      get copyWith => __$$GetCategoriesSuccessImplCopyWithImpl<T,
-          _$GetCategoriesSuccessImpl<T>>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return getCategoriesSuccess(data);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return getCategoriesSuccess?.call(data);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoriesSuccess != null) {
-      return getCategoriesSuccess(data);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return getCategoriesSuccess(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return getCategoriesSuccess?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoriesSuccess != null) {
-      return getCategoriesSuccess(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class GetCategoriesSuccess<T> implements HomeState<T> {
-  const factory GetCategoriesSuccess(final T data) =
-      _$GetCategoriesSuccessImpl<T>;
-
-  T get data;
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$GetCategoriesSuccessImplCopyWith<T, _$GetCategoriesSuccessImpl<T>>
-      get copyWith => throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class _$$GetCategoriesFailureImplCopyWith<T, $Res> {
-  factory _$$GetCategoriesFailureImplCopyWith(
-          _$GetCategoriesFailureImpl<T> value,
-          $Res Function(_$GetCategoriesFailureImpl<T>) then) =
-      __$$GetCategoriesFailureImplCopyWithImpl<T, $Res>;
-  @useResult
-  $Res call({String errorMessage});
-}
-
-/// @nodoc
-class __$$GetCategoriesFailureImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$GetCategoriesFailureImpl<T>>
-    implements _$$GetCategoriesFailureImplCopyWith<T, $Res> {
-  __$$GetCategoriesFailureImplCopyWithImpl(_$GetCategoriesFailureImpl<T> _value,
-      $Res Function(_$GetCategoriesFailureImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? errorMessage = null,
-  }) {
-    return _then(_$GetCategoriesFailureImpl<T>(
-      null == errorMessage
-          ? _value.errorMessage
-          : errorMessage // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$GetCategoriesFailureImpl<T> implements GetCategoriesFailure<T> {
-  const _$GetCategoriesFailureImpl(this.errorMessage);
-
-  @override
-  final String errorMessage;
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.getCategoriesFailure(errorMessage: $errorMessage)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$GetCategoriesFailureImpl<T> &&
-            (identical(other.errorMessage, errorMessage) ||
-                other.errorMessage == errorMessage));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, errorMessage);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$GetCategoriesFailureImplCopyWith<T, _$GetCategoriesFailureImpl<T>>
-      get copyWith => __$$GetCategoriesFailureImplCopyWithImpl<T,
-          _$GetCategoriesFailureImpl<T>>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return getCategoriesFailure(errorMessage);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return getCategoriesFailure?.call(errorMessage);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoriesFailure != null) {
-      return getCategoriesFailure(errorMessage);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return getCategoriesFailure(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return getCategoriesFailure?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoriesFailure != null) {
-      return getCategoriesFailure(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class GetCategoriesFailure<T> implements HomeState<T> {
-  const factory GetCategoriesFailure(final String errorMessage) =
-      _$GetCategoriesFailureImpl<T>;
-
-  String get errorMessage;
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$GetCategoriesFailureImplCopyWith<T, _$GetCategoriesFailureImpl<T>>
-      get copyWith => throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class _$$GetCategoryProductLoadingImplCopyWith<T, $Res> {
-  factory _$$GetCategoryProductLoadingImplCopyWith(
-          _$GetCategoryProductLoadingImpl<T> value,
-          $Res Function(_$GetCategoryProductLoadingImpl<T>) then) =
-      __$$GetCategoryProductLoadingImplCopyWithImpl<T, $Res>;
-}
-
-/// @nodoc
-class __$$GetCategoryProductLoadingImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$GetCategoryProductLoadingImpl<T>>
-    implements _$$GetCategoryProductLoadingImplCopyWith<T, $Res> {
-  __$$GetCategoryProductLoadingImplCopyWithImpl(
-      _$GetCategoryProductLoadingImpl<T> _value,
-      $Res Function(_$GetCategoryProductLoadingImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-}
-
-/// @nodoc
-
-class _$GetCategoryProductLoadingImpl<T>
-    implements GetCategoryProductLoading<T> {
-  const _$GetCategoryProductLoadingImpl();
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.getCategoryProductsLoading()';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$GetCategoryProductLoadingImpl<T>);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return getCategoryProductsLoading();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return getCategoryProductsLoading?.call();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoryProductsLoading != null) {
-      return getCategoryProductsLoading();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return getCategoryProductsLoading(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return getCategoryProductsLoading?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoryProductsLoading != null) {
-      return getCategoryProductsLoading(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class GetCategoryProductLoading<T> implements HomeState<T> {
-  const factory GetCategoryProductLoading() =
-      _$GetCategoryProductLoadingImpl<T>;
-}
-
-/// @nodoc
-abstract class _$$GetCategoryProductSuccessImplCopyWith<T, $Res> {
-  factory _$$GetCategoryProductSuccessImplCopyWith(
-          _$GetCategoryProductSuccessImpl<T> value,
-          $Res Function(_$GetCategoryProductSuccessImpl<T>) then) =
-      __$$GetCategoryProductSuccessImplCopyWithImpl<T, $Res>;
-  @useResult
-  $Res call({T data});
-}
-
-/// @nodoc
-class __$$GetCategoryProductSuccessImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$GetCategoryProductSuccessImpl<T>>
-    implements _$$GetCategoryProductSuccessImplCopyWith<T, $Res> {
-  __$$GetCategoryProductSuccessImplCopyWithImpl(
-      _$GetCategoryProductSuccessImpl<T> _value,
-      $Res Function(_$GetCategoryProductSuccessImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? data = freezed,
-  }) {
-    return _then(_$GetCategoryProductSuccessImpl<T>(
-      freezed == data
-          ? _value.data
-          : data // ignore: cast_nullable_to_non_nullable
-              as T,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$GetCategoryProductSuccessImpl<T>
-    implements GetCategoryProductSuccess<T> {
-  const _$GetCategoryProductSuccessImpl(this.data);
-
-  @override
-  final T data;
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.getCategoryProductsSuccess(data: $data)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$GetCategoryProductSuccessImpl<T> &&
-            const DeepCollectionEquality().equals(other.data, data));
-  }
-
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, const DeepCollectionEquality().hash(data));
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$GetCategoryProductSuccessImplCopyWith<T,
-          _$GetCategoryProductSuccessImpl<T>>
-      get copyWith => __$$GetCategoryProductSuccessImplCopyWithImpl<T,
-          _$GetCategoryProductSuccessImpl<T>>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return getCategoryProductsSuccess(data);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return getCategoryProductsSuccess?.call(data);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoryProductsSuccess != null) {
-      return getCategoryProductsSuccess(data);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return getCategoryProductsSuccess(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return getCategoryProductsSuccess?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoryProductsSuccess != null) {
-      return getCategoryProductsSuccess(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class GetCategoryProductSuccess<T> implements HomeState<T> {
-  const factory GetCategoryProductSuccess(final T data) =
-      _$GetCategoryProductSuccessImpl<T>;
-
-  T get data;
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$GetCategoryProductSuccessImplCopyWith<T,
-          _$GetCategoryProductSuccessImpl<T>>
-      get copyWith => throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class _$$GetCategoryProductFailureImplCopyWith<T, $Res> {
-  factory _$$GetCategoryProductFailureImplCopyWith(
-          _$GetCategoryProductFailureImpl<T> value,
-          $Res Function(_$GetCategoryProductFailureImpl<T>) then) =
-      __$$GetCategoryProductFailureImplCopyWithImpl<T, $Res>;
-  @useResult
-  $Res call({String errorMessage});
-}
-
-/// @nodoc
-class __$$GetCategoryProductFailureImplCopyWithImpl<T, $Res>
-    extends _$HomeStateCopyWithImpl<T, $Res, _$GetCategoryProductFailureImpl<T>>
-    implements _$$GetCategoryProductFailureImplCopyWith<T, $Res> {
-  __$$GetCategoryProductFailureImplCopyWithImpl(
-      _$GetCategoryProductFailureImpl<T> _value,
-      $Res Function(_$GetCategoryProductFailureImpl<T>) _then)
-      : super(_value, _then);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? errorMessage = null,
-  }) {
-    return _then(_$GetCategoryProductFailureImpl<T>(
-      null == errorMessage
-          ? _value.errorMessage
-          : errorMessage // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$GetCategoryProductFailureImpl<T>
-    implements GetCategoryProductFailure<T> {
-  const _$GetCategoryProductFailureImpl(this.errorMessage);
-
-  @override
-  final String errorMessage;
-
-  @override
-  String toString() {
-    return 'HomeState<$T>.getCategoryProductsFailure(errorMessage: $errorMessage)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$GetCategoryProductFailureImpl<T> &&
-            (identical(other.errorMessage, errorMessage) ||
-                other.errorMessage == errorMessage));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, errorMessage);
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$GetCategoryProductFailureImplCopyWith<T,
-          _$GetCategoryProductFailureImpl<T>>
-      get copyWith => __$$GetCategoryProductFailureImplCopyWithImpl<T,
-          _$GetCategoryProductFailureImpl<T>>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
-    required TResult Function() getHomeProductsLoading,
-    required TResult Function(T data) getHomeProductsSuccess,
-    required TResult Function(String errorMessage) getHomeProductsFailure,
-  }) {
-    return getCategoryProductsFailure(errorMessage);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
-    TResult? Function()? getHomeProductsLoading,
-    TResult? Function(T data)? getHomeProductsSuccess,
-    TResult? Function(String errorMessage)? getHomeProductsFailure,
-  }) {
-    return getCategoryProductsFailure?.call(errorMessage);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
-    TResult Function()? getHomeProductsLoading,
-    TResult Function(T data)? getHomeProductsSuccess,
-    TResult Function(String errorMessage)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoryProductsFailure != null) {
-      return getCategoryProductsFailure(errorMessage);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
-    required TResult Function(GetHomeProductLoading<T> value)
-        getHomeProductsLoading,
-    required TResult Function(GetHomeProductSuccess<T> value)
-        getHomeProductsSuccess,
-    required TResult Function(GetHomeProductFailure<T> value)
-        getHomeProductsFailure,
-  }) {
-    return getCategoryProductsFailure(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-  }) {
-    return getCategoryProductsFailure?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
-    TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
-    TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
-    TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
-    required TResult orElse(),
-  }) {
-    if (getCategoryProductsFailure != null) {
-      return getCategoryProductsFailure(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class GetCategoryProductFailure<T> implements HomeState<T> {
-  const factory GetCategoryProductFailure(final String errorMessage) =
-      _$GetCategoryProductFailureImpl<T>;
-
-  String get errorMessage;
-
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$GetCategoryProductFailureImplCopyWith<T,
-          _$GetCategoryProductFailureImpl<T>>
-      get copyWith => throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -2309,15 +257,6 @@ class _$GetHomeProductLoadingImpl<T> implements GetHomeProductLoading<T> {
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
     required TResult Function() getHomeProductsLoading,
     required TResult Function(T data) getHomeProductsSuccess,
     required TResult Function(String errorMessage) getHomeProductsFailure,
@@ -2329,15 +268,6 @@ class _$GetHomeProductLoadingImpl<T> implements GetHomeProductLoading<T> {
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
     TResult? Function()? getHomeProductsLoading,
     TResult? Function(T data)? getHomeProductsSuccess,
     TResult? Function(String errorMessage)? getHomeProductsFailure,
@@ -2349,15 +279,6 @@ class _$GetHomeProductLoadingImpl<T> implements GetHomeProductLoading<T> {
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
     TResult Function()? getHomeProductsLoading,
     TResult Function(T data)? getHomeProductsSuccess,
     TResult Function(String errorMessage)? getHomeProductsFailure,
@@ -2373,21 +294,6 @@ class _$GetHomeProductLoadingImpl<T> implements GetHomeProductLoading<T> {
   @optionalTypeArgs
   TResult map<TResult extends Object?>({
     required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
     required TResult Function(GetHomeProductLoading<T> value)
         getHomeProductsLoading,
     required TResult Function(GetHomeProductSuccess<T> value)
@@ -2402,18 +308,6 @@ class _$GetHomeProductLoadingImpl<T> implements GetHomeProductLoading<T> {
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>({
     TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -2425,18 +319,6 @@ class _$GetHomeProductLoadingImpl<T> implements GetHomeProductLoading<T> {
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -2526,15 +408,6 @@ class _$GetHomeProductSuccessImpl<T> implements GetHomeProductSuccess<T> {
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
     required TResult Function() getHomeProductsLoading,
     required TResult Function(T data) getHomeProductsSuccess,
     required TResult Function(String errorMessage) getHomeProductsFailure,
@@ -2546,15 +419,6 @@ class _$GetHomeProductSuccessImpl<T> implements GetHomeProductSuccess<T> {
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
     TResult? Function()? getHomeProductsLoading,
     TResult? Function(T data)? getHomeProductsSuccess,
     TResult? Function(String errorMessage)? getHomeProductsFailure,
@@ -2566,15 +430,6 @@ class _$GetHomeProductSuccessImpl<T> implements GetHomeProductSuccess<T> {
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
     TResult Function()? getHomeProductsLoading,
     TResult Function(T data)? getHomeProductsSuccess,
     TResult Function(String errorMessage)? getHomeProductsFailure,
@@ -2590,21 +445,6 @@ class _$GetHomeProductSuccessImpl<T> implements GetHomeProductSuccess<T> {
   @optionalTypeArgs
   TResult map<TResult extends Object?>({
     required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
     required TResult Function(GetHomeProductLoading<T> value)
         getHomeProductsLoading,
     required TResult Function(GetHomeProductSuccess<T> value)
@@ -2619,18 +459,6 @@ class _$GetHomeProductSuccessImpl<T> implements GetHomeProductSuccess<T> {
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>({
     TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -2642,18 +470,6 @@ class _$GetHomeProductSuccessImpl<T> implements GetHomeProductSuccess<T> {
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -2752,15 +568,6 @@ class _$GetHomeProductFailureImpl<T> implements GetHomeProductFailure<T> {
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() initial,
-    required TResult Function() sliderLoading,
-    required TResult Function(T data) sliderSuccess,
-    required TResult Function(String errorMessage) sliderFailure,
-    required TResult Function() getCategoriesLoading,
-    required TResult Function(T data) getCategoriesSuccess,
-    required TResult Function(String errorMessage) getCategoriesFailure,
-    required TResult Function() getCategoryProductsLoading,
-    required TResult Function(T data) getCategoryProductsSuccess,
-    required TResult Function(String errorMessage) getCategoryProductsFailure,
     required TResult Function() getHomeProductsLoading,
     required TResult Function(T data) getHomeProductsSuccess,
     required TResult Function(String errorMessage) getHomeProductsFailure,
@@ -2772,15 +579,6 @@ class _$GetHomeProductFailureImpl<T> implements GetHomeProductFailure<T> {
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult? Function()? initial,
-    TResult? Function()? sliderLoading,
-    TResult? Function(T data)? sliderSuccess,
-    TResult? Function(String errorMessage)? sliderFailure,
-    TResult? Function()? getCategoriesLoading,
-    TResult? Function(T data)? getCategoriesSuccess,
-    TResult? Function(String errorMessage)? getCategoriesFailure,
-    TResult? Function()? getCategoryProductsLoading,
-    TResult? Function(T data)? getCategoryProductsSuccess,
-    TResult? Function(String errorMessage)? getCategoryProductsFailure,
     TResult? Function()? getHomeProductsLoading,
     TResult? Function(T data)? getHomeProductsSuccess,
     TResult? Function(String errorMessage)? getHomeProductsFailure,
@@ -2792,15 +590,6 @@ class _$GetHomeProductFailureImpl<T> implements GetHomeProductFailure<T> {
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? initial,
-    TResult Function()? sliderLoading,
-    TResult Function(T data)? sliderSuccess,
-    TResult Function(String errorMessage)? sliderFailure,
-    TResult Function()? getCategoriesLoading,
-    TResult Function(T data)? getCategoriesSuccess,
-    TResult Function(String errorMessage)? getCategoriesFailure,
-    TResult Function()? getCategoryProductsLoading,
-    TResult Function(T data)? getCategoryProductsSuccess,
-    TResult Function(String errorMessage)? getCategoryProductsFailure,
     TResult Function()? getHomeProductsLoading,
     TResult Function(T data)? getHomeProductsSuccess,
     TResult Function(String errorMessage)? getHomeProductsFailure,
@@ -2816,21 +605,6 @@ class _$GetHomeProductFailureImpl<T> implements GetHomeProductFailure<T> {
   @optionalTypeArgs
   TResult map<TResult extends Object?>({
     required TResult Function(HomeInitial<T> value) initial,
-    required TResult Function(SliderLoading<T> value) sliderLoading,
-    required TResult Function(SliderSuccess<T> value) sliderSuccess,
-    required TResult Function(SliderFailure<T> value) sliderFailure,
-    required TResult Function(GetCategoriesLoading<T> value)
-        getCategoriesLoading,
-    required TResult Function(GetCategoriesSuccess<T> value)
-        getCategoriesSuccess,
-    required TResult Function(GetCategoriesFailure<T> value)
-        getCategoriesFailure,
-    required TResult Function(GetCategoryProductLoading<T> value)
-        getCategoryProductsLoading,
-    required TResult Function(GetCategoryProductSuccess<T> value)
-        getCategoryProductsSuccess,
-    required TResult Function(GetCategoryProductFailure<T> value)
-        getCategoryProductsFailure,
     required TResult Function(GetHomeProductLoading<T> value)
         getHomeProductsLoading,
     required TResult Function(GetHomeProductSuccess<T> value)
@@ -2845,18 +619,6 @@ class _$GetHomeProductFailureImpl<T> implements GetHomeProductFailure<T> {
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>({
     TResult? Function(HomeInitial<T> value)? initial,
-    TResult? Function(SliderLoading<T> value)? sliderLoading,
-    TResult? Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult? Function(SliderFailure<T> value)? sliderFailure,
-    TResult? Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult? Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult? Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult? Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult? Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult? Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult? Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult? Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult? Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
@@ -2868,18 +630,6 @@ class _$GetHomeProductFailureImpl<T> implements GetHomeProductFailure<T> {
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(HomeInitial<T> value)? initial,
-    TResult Function(SliderLoading<T> value)? sliderLoading,
-    TResult Function(SliderSuccess<T> value)? sliderSuccess,
-    TResult Function(SliderFailure<T> value)? sliderFailure,
-    TResult Function(GetCategoriesLoading<T> value)? getCategoriesLoading,
-    TResult Function(GetCategoriesSuccess<T> value)? getCategoriesSuccess,
-    TResult Function(GetCategoriesFailure<T> value)? getCategoriesFailure,
-    TResult Function(GetCategoryProductLoading<T> value)?
-        getCategoryProductsLoading,
-    TResult Function(GetCategoryProductSuccess<T> value)?
-        getCategoryProductsSuccess,
-    TResult Function(GetCategoryProductFailure<T> value)?
-        getCategoryProductsFailure,
     TResult Function(GetHomeProductLoading<T> value)? getHomeProductsLoading,
     TResult Function(GetHomeProductSuccess<T> value)? getHomeProductsSuccess,
     TResult Function(GetHomeProductFailure<T> value)? getHomeProductsFailure,
