+++ lib/features/home/presentation/widgets/category_product_bloc_builder.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:thimar/core/theming/app_colors.dart';
-import 'package:thimar/features/home/logic/home_cubit.dart';
-import 'package:thimar/features/home/logic/home_state.dart';
+import 'package:thimar/features/home/logic/category_cubit.dart';
+import 'package:thimar/features/home/logic/category_state.dart';
 
 import 'category_product_grid_view.dart';
 
@@ -11,28 +11,34 @@ class CategoryProductBlocBuilder extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<HomeCubit, HomeState>(builder: (context, state) {
-      return state.maybeWhen(
-        getCategoryProductsLoading: () =>  Center(
-          child: CircularProgressIndicator(
-            color: AppColors.primaryColor,
-          ),
-        ),
-        getCategoryProductsSuccess: (data) => CategoryProductGridView(
-          data: data,
-        ),
-        getCategoryProductsFailure: (error) => Center(
-          child: Text(
-            error,
-            style: TextStyle(
-              color: Colors.red,
-              fontSize: 16,
-              fontWeight: FontWeight.w400,
+    return BlocBuilder<CategoryCubit, CategoryState>(
+        buildWhen: (previous, current) =>
+            previous != current &&
+            (current is GetCategoryProductLoading ||
+                current is GetCategoryProductSuccess ||
+                current is GetCategoryProductFailure),
+        builder: (context, state) {
+          return state.maybeWhen(
+            getCategoryProductsLoading: () => Center(
+              child: CircularProgressIndicator(
+                color: AppColors.primaryColor,
+              ),
             ),
-          ),
-        ),
-        orElse: () => const SizedBox(),
-      );
-    });
+            getCategoryProductsSuccess: (data) => CategoryProductGridView(
+              data: data,
+            ),
+            getCategoryProductsFailure: (error) => Center(
+              child: Text(
+                error,
+                style: TextStyle(
+                  color: Colors.red,
+                  fontSize: 16,
+                  fontWeight: FontWeight.w400,
+                ),
+              ),
+            ),
+            orElse: () => const SizedBox(),
+          );
+        });
   }
 }
